Add files via upload
Using smtplib to send mail to a specified email id should the price of a particular product on amazon fall below a certain value.
In order for this to work, beautifulsoup and smtplib are prerequisites. 
Google app password needs to be generated as well.
URL should contain the address of the article whose price we want to keep a watch on.
The program will run in the background and check the price every 2 hours(can be changed) and notify the user by sending mail at specified address should the price fall below a certain value.